requirements.txt

# Core ML and Data Processing
# numpy==1.24.3
# pandas==2.0.3
# scikit-learn==1.3.0
xgboost==1.7.6

joblib==1.3.2

# API Backend - Using compatible versions
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# NLP and RAG - Simplified for now to avoid conflicts
transformers==4.36.2
sentence-transformers==2.2.2
torch==2.1.2

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
pytest==7.4.4


